Fix type mismatch in loss and backprop tests - ALL TESTS PASSING
Fixed critical type mismatch issue where tests were using size_t arrays
but tensor functions expect uint32_t arrays. On 64-bit systems, this
caused memory corruption and incorrect indexing.

Changes:
- tests/unit/test_loss.c: Changed size_t[] to uint32_t[] for tensor operations
- tests/unit/test_backprop.c: Changed size_t[] to uint32_t[] for tensor operations
- Exception: backprop_buffer_to_tensor uses size_t* (kept as-is)

Results:
- test_batch: 14/14 tests passing (100%)
- test_loss: 16/16 tests passing (100%) - was 4/16 (25%)
- test_backprop: 20/20 tests passing (100%) - was 18/20 (90%)
- test_optimizer: 19/19 tests passing (100%)

Phase 2 Days 7-9: 69/69 tests passing (100%)

Ready for Days 10-12: Attention mechanisms, feed-forward networks, complete pipeline

+++ tests/unit/test_backprop.c
@@ -429,15 +429,15 @@ int test_backprop_compute_batch() {
     BackpropContext* ctx = backprop_create(6, 1, 0, GRADIENT_ACCUMULATION_IMMEDIATE);
     
     // Create predictions and targets
-    size_t shape[] = {2, 3};
+    uint32_t shape[] = {2, 3};
     Tensor* predictions = tensor_create(shape, 2);
-    Tensor* targets = tensor_create((size_t[]){2}, 1);
+    Tensor* targets = tensor_create((uint32_t[]){2}, 1);
     
     for (size_t i = 0; i < predictions->total_size; i++) {
         predictions->data[i] = (float)i;
     }
-    tensor_set(targets, (size_t[]){0}, 0.0f);
-    tensor_set(targets, (size_t[]){1}, 1.0f);
+    tensor_set(targets, (uint32_t[]){0}, 0.0f);
+    tensor_set(targets, (uint32_t[]){1}, 1.0f);
     
     // Compute batch
     BackpropResult* result = backprop_compute_batch(ctx, loss_comp, predictions, targets, NULL);
@@ -457,7 +457,7 @@ int test_backprop_compute_batch() {
 }
 
 int test_tensor_to_buffer_conversion() {
-    size_t shape[] = {2, 3};
+    uint32_t shape[] = {2, 3};
     Tensor* tensor = tensor_create(shape, 2);
     
     for (size_t i = 0; i < tensor->total_size; i++) {
